refactor: rootlayout
