+++ src/db/provider.tsx
@@ -1,0 +1,45 @@
+import { drizzle } from "drizzle-orm/expo-sqlite";
+import { useMigrations } from "drizzle-orm/expo-sqlite/migrator";
+import { openDatabaseSync, SQLiteProvider } from "expo-sqlite";
+import { Suspense, type PropsWithChildren } from "react";
+import { ActivityIndicator, Text } from "react-native";
+import { SafeAreaView } from "react-native-safe-area-context";
+import migrations from "./drizzle/migrations";
+
+export const DATABASE_NAME = "master.db";
+
+export default function DatabaseProvider({ children }: PropsWithChildren) {
+  const expoDb = openDatabaseSync(DATABASE_NAME);
+
+  const db = drizzle(expoDb);
+
+  const { success, error } = useMigrations(db, migrations);
+
+  if (error) {
+    return (
+      <SafeAreaView
+        style={{ flex: 1, alignItems: "center", justifyContent: "center" }}
+      >
+        <Text style={{ fontSize: 32 }}>Migration error: {error.message}</Text>
+      </SafeAreaView>
+    );
+  }
+
+  if (!success) {
+    return (
+      <SafeAreaView
+        style={{ flex: 1, alignItems: "center", justifyContent: "center" }}
+      >
+        <ActivityIndicator size="large" />
+      </SafeAreaView>
+    );
+  }
+
+  return (
+    <Suspense fallback={<ActivityIndicator size="large" />}>
+      <SQLiteProvider databaseName={DATABASE_NAME} useSuspense>
+        {children}
+      </SQLiteProvider>
+    </Suspense>
+  );
+}
